organise tests by type
